+++ src/main/java/io/xpring/payid/PayIDClient.java
@@ -19,8 +19,11 @@ import java.util.Map;
 /**
  * Implements interaction with a PayID service.
  * Warning:  This class is experimental and should not be used in production applications.
+ *
+ * @deprecated Use the idiomatically named `PayIdClient` class instead.
  */
 @SuppressWarnings("checkstyle:AbbreviationAsWordInName")
+@Deprecated
 public class PayIDClient {
   /**
    * The version of PayID.