+++ src/main/java/io/xpring/payid/idiomatic/XrpPayIdClient.java
@@ -1,6 +1,6 @@
-package io.xpring.payid;
+package io.xpring.payid.idiomatic;
 
-import io.xpring.common.XRPLNetwork;
+import io.xpring.common.idiomatic.XrplNetwork;
 import io.xpring.payid.generated.model.CryptoAddressDetails;
 import io.xpring.xrpl.ClassicAddress;
 import io.xpring.xrpl.ImmutableClassicAddress;
@@ -8,33 +8,30 @@ import io.xpring.xrpl.Utils;
 
 /**
  * Provides functionality for XRP in the PayID protocol.
- *
- * @deprecated Use the idiomatically named `XrpPayIdClient` class instead.
  */
-@SuppressWarnings("checkstyle:AbbreviationAsWordInName")
-@Deprecated
-public class XRPPayIDClient extends PayIDClient implements XRPPayIDClientInterface {
+public class XrpPayIdClient extends PayIdClient implements XrpPayIdClientInterface {
   /**
    * The XRP Ledger network that this client attaches to.
    */
-  private XRPLNetwork xrplNetwork;
+  private XrplNetwork xrplNetwork;
 
   /**
    * Get the XRP Ledger network that this client attaches to.
    *
    * @return The XRP Ledger network that this client attaches to.
    */
-  public XRPLNetwork getXRPLNetwork() {
+  public XrplNetwork getXrplNetwork() {
     return this.xrplNetwork;
   }
 
   /**
-   * Construct a new XRPPayIDClient.
+   * Construct a new XrpPayIdClient.
    *
    * @param xrplNetwork The XRP Ledger network that this client attaches to.
    */
-  public XRPPayIDClient(XRPLNetwork xrplNetwork) {
-    super();
+  public XrpPayIdClient(XrplNetwork xrplNetwork) {
+    super("xrpl-" + xrplNetwork.getNetworkName());
+
     this.xrplNetwork = xrplNetwork;
   }
 
@@ -46,12 +43,12 @@ public class XRPPayIDClient extends PayIDClient implements XRPPayIDClientInterfa
    * </p>
    * @see "https://xrpaddress.info"
    *
-   * @param payID The payID to resolve for an address.
+   * @param payId The PayID to resolve for an address.
    * @return An XRP address representing the given PayID.
-   * @throws PayIDException if the inputs were invalid.
+   * @throws PayIdException if the inputs were invalid.
    */
-  public String xrpAddressForPayID(String payID) throws PayIDException {
-    CryptoAddressDetails addressDetails = super.cryptoAddressForPayId(payID, this.xrplNetwork.getNetworkName());
+  public String xrpAddressForPayId(String payId) throws PayIdException {
+    CryptoAddressDetails addressDetails = super.addressForPayId(payId);
 
     // Return address immediately if it is an X-Address.
     String address = addressDetails.getAddress();
@@ -60,7 +57,7 @@ public class XRPPayIDClient extends PayIDClient implements XRPPayIDClientInterfa
     }
 
     // Otherwise, build a classic address.
-    boolean isTest = this.xrplNetwork != XRPLNetwork.MAIN;
+    boolean isTest = this.xrplNetwork != XrplNetwork.MAIN;
     ImmutableClassicAddress.Builder classicAddressBuilder = ImmutableClassicAddress.builder()
         .address(address)
         .isTest(isTest);
@@ -73,8 +70,8 @@ public class XRPPayIDClient extends PayIDClient implements XRPPayIDClientInterfa
     // Encode and return the classic address to an X-Address.
     String encodedXAddress = Utils.encodeXAddress(classicAddress);
     if (encodedXAddress == null) {
-      throw new PayIDException(
-          PayIDExceptionType.UNEXPECTED_RESPONSE,
+      throw new PayIdException(
+          PayIdExceptionType.UNEXPECTED_RESPONSE,
           "The returned address was in an unexpected format"
       );
     }